refactor(authentication): leverage Express req/res types
Now that the REST layer uses Express request/response types internally,
we can simplify StrategyAdapter by removing ShimRequest.

While making these changes, other places are cleaned up as well.
