+++ packages/authentication/src/index.ts
@@ -3,10 +3,8 @@
 // This file is licensed under the MIT License.
 // License text available at https://opensource.org/licenses/MIT
 
+export * from './types';
 export * from './authentication.component';
 export * from './decorators';
 export * from './keys';
 export * from './strategy-adapter';
-
-// internals for tests
-export * from './providers';
